👨‍🔧 Cli için yapılandırma tamamlandı
